Begin work to convert the editor.js file to a module.
